iop: Consistency checks when (un)registering packaes.
Following I507913ffd, we can get in a situation where some classes are
registered while their parents are not (because some explicit register are
missing) which can be dangerous.

Add a full-check at the end of iop_register_packages_env /
iop_unregister_packages to ensure this is not the case.

Changes will be needed in platform and mmsx following this patch.

Change-Id: I7d2042da50479893b701798fb67457862a01de35