iop: Fix loading of multiple DSOs following previous commits.
The consistency check added in 538e22c1512a broke the loading of multiple
DSOs (using python for example). Indeed, this is possible in that case to
have several different packages having the same name, and we had the error
"a package named 'XXX' was already registered".

In order to properly fix that:
* have multiple packages/enums as values of the pkgs_by_name /
  enums_by_fullname hash tables of the environment, since they are not
  necessarily unique.
* Add a flag IOP_REGPKG_FROM_DSO at packages registration, to know if this
  is a package from a DSO (and in that case allow duplicates).

Change-Id: I281ecd63d1d0bb2dc65516891e6a904a1190d5ec