core-obj: ensure fields order in classes
In core obj classes, we expect the fields to be declared from the top
super class to the last inheriting class so instances of each class can be
casted into instances of any of their super class.

Messing up with fields order when declaring classes can end up in pretty
nasty segmentation faults.

Example:
~~~~~~~~
    /* What we have. */
    #define MY_CLASS_FIELDS(pfx)                \
        uint32_t foo;                           \
        OBJECT_FIELDS(pfx)

    /* What we should have. */
    #define MY_CLASS_FIELDS(pfx)                \
        OBJECT_FIELDS(pfx);                     \
        uint32_t foo

In that case, it can be perfectly OK to create an object with obj_new(),
but the whole object will be messed up at the time of the cast, leading to
segmentation faults that can be hard to understand.

This patch adds a STATIC_ASSERT that will prevent the code from compiling
when the fields order in the class is incorrect.

Change-Id: If372cf2d71083c14f6fd5920d0997a525c75584d
Signed-off-by: Sébastien Thouément <sebastien.thouement@intersec.com>