Homogenize *.in.c files
Before this patch, there were some *-inl.c files, some *.inc.c files and
some *.c for which it was hard to understand quickly that these files
couldn't actually be compiled by themselves and must be included in
another files that declare some context. This patch homogenizes this in
order, amongst others to allow some tools such as clang-tidy to exclude
files that only make sense when they are included in other files.

Change-Id: If0b2e0143ea617da836eacd8422ab8a0969587ce