core-mem-stack: update t_fmt to make it use t_pool() function
Before this patch, t_fmt used '&t_pool_g.funcs' which is supposed to be
internal, moreover there is the function t_pool() that does exactly the
same thing: returning &t_pool_g.funcs.

This patch is essentially needed to ease the handling and writing of rules
about the t_pool in tools such as clang-tidy: if we want to catch uses of
the t_pool, it is much simple to only have to match functions with name
starting with 't_'. This patch allows us to avoid a special handling for
t_fmt since now it simply calls a 't_*' function, which is caught by the
general rule.

Change-Id: I3f3ae80f46382e1a906dfd5628c06e53dafe6784