core: rework the rand() stuff.
We used to use the HAVEGE random number generator, however that
implementation proved to be extremely inefficient since it must recompute
8192 number every time the previous 8192 have been consumed and that
computation is pretty expensive. In the end, the generator is between 6
and 30 times slower than rand() (depending on whether we hit a "refill")
for a globally equivalent randomess (see the random bitmap generated from
ha_rand()[1] and rand()[2])

Moreover, the generator was not thread safe, and since it had to maintain
a large pool of data it would have been quite expensive to make it
thread-local.

As a consequence, we completely get rid of HAVEGE and replace it with a
thin wrapper around rand_r(). The libcommon library now provides a rand()
implementation that is thread-safe since the module hides the thread-local
seed. Helpers for generating 32-bit and 64-bit integers with guarantee
that all bits are part of the random number are also added, and all the
ha_* based function are replaced by new variants.

[1] https://img.corp/RIyZPdJt
[2] https://img.corp/tCpxBavy

Change-Id: If0764248b9afb2be7aa5e41846636381f6ce230b
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>