log: Introduce IS_LOG_TIMESTAMP.
If this environment variable is set to a positive numeric value, the
default log handlers are now printing a timestamp (with a 10^-2 sec
precision).

Here are some examples of the output with the standard handler:

  1453375240.22 zpf.master@1[5431]: platform/master.blk:4586:__platform_start_master_block_func_0: from e_trace
  1453375240.22 zpf.master@1[5431]: platform/master.blk:4587:__platform_start_master_block_func_0: {platform/master} from logger_trace
  1453375240.22 zpf.master@1[5431]: warn:  {platform/master} from logger

...and with the fancy handler:

  1453375312.78 __platform_sta...: from e_trace                                platform/master.blk:4586:zpf.master@1[6249]
  1453375312.78 __platform_sta...: {platform/master} from logger_trace         platform/master.blk:4587:zpf.master@1[6249]
  1453375312.78 zpf.master@1[6249]: {platform/master} from logger

The goal of this patch is to activate IS_LOG_TIMESTAMP on the buildbots,
in order to ease debugging in case of test failure.

Change-Id: I5cd45fa38070939f46e7ae034441de316556e676