iopc: fix leaks of iopc_g.loc_stack
All iopc modules uses the global variable iopc_g.loc_stack to print
informative logs with location information. This variable is a vector that
is never wiped.

To fix this leak, a module "iopc" is added that is responsible for
initializing and wiping this vector. This module is added in
iopc-typer.blk and not iopc.c because iopc.c is the code of the
executable "iopc" and not linked in libiopc.

Change-Id: I9e141f1b1cb60225cc1535a5d74fe8cb67da4fb2