iop-rpc-channel: fix handling of truncated unix messages.
Since 484ef72, we wrongly check that there is no 'fd overflow' in case we
didn't receive all the expected fd from a unix ichannel. Not receiving a
fd can happen in case there is no free file descriptor on our side of the
channel. In that particular case, the recvmsg() call returns successfully,
but fill errno with EMFILE and set the flag MSG_CTRUNC.

Change-Id: Ib5c7f4b23be503682f918374e3adac71ae1e41cf
Refs: #42961
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>