unix: backup errno in panic handler.
Since the panic handler performs a few syscalls, it can alter the value of
errno if anything goes wrong during the execution of one of them
(open/write/...). We now backup and restore the previous value of errno in
order to be able to get the value that was set during the execution of the
operation that triggered the panic.

Change-Id: Ie4f10fa778bf72c67c8d41bb49c2018727ea87a1
Refs: #42961
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>