net-addr: port addr_filter_build() from mcg/connector-main.blk
This function is used to build a addr_filter_t from a subnetwork.

A subnetwork is identified by its CIDR notation, e.g. 192.168.0.1/24 or
fe80::202:b3ff:fe1e:8329/32, or by a single IP address, 192.168.0.12 or
ff09::1234:abcd.

It is a port of the function addr_filters_build() from MCG with the fix
I5e3b592adc.

Change-Id: Id54f41e2656544760eb984b20dc76a878d1f8422
Refs: #42693