qps: Use logger_fatal instead of exit/_exit.
The origin of this patch is a behavior observed by Brice on its machine:
his statd segfaulted in a destructor after the call of exit in qps_enospc.

I could have changed 'exit' to '_exit' to avoid that (which would have
been consistent with the call just above), but in products life, the logs
emitted just beore a exit/_exit are totally lost (because not sent to the
logd), and using logger_fatal avoids that because the message is written
in the machine syslogs.

The drawback is that we lose the exit code QPS_EX_ENOSPC, but it was
probably not used at all.

Change-Id: Idcc539ec8fb5bb724d20821dc70a6bbe86b6467e