iopc[json]: Dump RPCs in the order of the IOP, not the order of the tags.
When generating a iop.json file, the order of the functions is now the
order of the IOP file, and not the order of the tags.

In order to not break the order of the generated C code, also perform the
by-tag sort in iopc-lang-c.

This will allow to generate more consistent documentations.

Change-Id: I6fb5a59a99d0782d05ccc4d4613bff118fe27141