log-file: add a flag to force file rotation when opening a log_file_t
Example of use case: when used to write results of a calculation to CSV
files, we might want to separate the results of different calculations
even if the file prefix is supposed to be the same. An other reason to
change file is that the header (if there is any) can be changed.

Change-Id: I87d0fef292c200cec5fec32aec28567d46b26812