iop-json: don't put too large integers in generated JSON.
Javascript does not support integers, only floating point numbers
(actually stored as doubles). As a consequence, it cannot correctly unpack
JSON files that contains integers that are not expressible as
(mantisse) * 2 ^ (exposant) where mantisse value is comprised in the range
[-(2^53-1);(2^53-1)]. As a consequence, it is quite generally accepted
that the safe range for integers in on doubles in that
[-(2^53-1);(2^53-1)] range. So if an integer is not in the safe range,
represent it as a string so we are sure the unpacking won't be
destructive.

Change-Id: I1c833383155ec4bae884d724f5e39db197aeb867
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>