iopc: pack large integer values as strings in JSON output.
Javascript does not supports integers, only floating point numbers
(actually stored as doubles). As a consequence, it cannot correctly unpack
JSON files that contains integers that are not expressible as
(mantisse) * 2 ^ (exposant) where mantisse value is comprised in the range
[-(2^53-1);(2^53-1)]. As a consequence, it is quite generally accepted
that the safe range for integers in on doubles in that
[-(2^53-1);(2^53-1)] range. So if an integer is not in the safe, range,
represent it as a string so we are sure the unpacking won't be
destructive.

Change-Id: Ic83639d0238085fd4e9f1290fe7d49841fd41737
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>