iop: fix binary unpacking with no longer existing tags
This is a backport of 18c8b7cdaf86099ddbfcd842ed8d1a74e38bdaea, the commit
message is kept for clarity:

When unpacking a class, tags that no longer exists in the struct but were
set could break the class unpacking.

When unpacking a class, a special tag "0" is used to separate the
different classes in the hierarchy, and allow finding the next class id.
However, in several cases, this tag "0" was expected but was raising an
error because tags of no longer existing fields were still in the buffer.

To solve this, unpack_struct has to do 2 cleanup verifications:
* consume every unknown tag (until the buffer is ended or the special tag
  is reached).
* check every existing tag that were not found in the unpacking.

Step 2 was correctly done, step 1 was not.

Change-Id: Ia2763c2a22565fd4650dafa9b367732b0c491e5f