xml: fix line information in error message
xmlTextReaderGetParserLineNumber and xmlTextReaderGetParserColumnNumber do
not return what we expect, it was known when these functions were
submitted:
https://mail.gnome.org/archives/xml/2005-January/msg00017.html

As said in this post, we can easily get the right line with xmlGetLineNo,
but not the column. It's still better than nothing, and we also have the
path to the current node.

Change-Id: I7f49e7230f522edc73e4df0903cf6977733ac372