iop: forbid private classes in public (un)packers.
The unpacker will fail if the FORBID_PRIVATE is set and a private class is
encountered.

The packer will assert in case SKIP_PRIVATE is provided and a private
class is found. We could have chosen to ignore the field if the field is
optional, but since the field is not private by itself, we don't expect it
to be automatically unset by a packer. As a consequence implementers of
public interface have to ensure their interface won't publish a private
class. In dev this will cause an assert, but in production this will just
be packed, causing emission of object that can't be sent back through the
public interface.

Change-Id: I0bba8f2edcc45427bc2be29b54a311f05d4d1db6
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>