container-qhash: introduce qh_seal/qm_seal
qhash_seal() puts the hash table in a "clean" state (no running resize, no
ghost entry) and protects the table against new modifications.

This function is costly and should be used only in contexts in which we
expect the number of elements in the table to be fixed at a time and
cannot use the iterative resize system during the find opeartions (for
example in multi-thread environments).

Change-Id: I5ee16c7fecd0dc5fd1f3902dcde1f723a3622e68