qhash: fix forgotten reset for ghosts count when clearing a table
The ghosts are deleted elements that need to be remembered in order to
keep the table working correctly. The need for this feature is related to
the way we handle collisions.

The current number of ghosts in the table is a metric used to choose
whether a resize is necessary or not, so the consequence of the forgotten
reset was that we could trigger a resize in conditions we would not have
been supposed to do so according to the resize rules.

Change-Id: If5836b5b4d227fab55d28a13cc66ac776d643a90
(cherry picked from commit 9cb14b993847ab3bd32ff2f8bd57f9add12b9e32)