ichannel: Fix incoming RPCs not being traced if not traced in the caller.
When receiving a RPC call, the ichannel layer were only relying on the
received flags to determine if the query should be traced or not.

This works fine for the internal queries, but do not work if we want to
trace queries coming from the website of from the python.

This commit fixes that by adding a registry of traced RPCs by command,
which should not hurt the performances too much.

Change-Id: I278210bbc5777be9cdf655ec67c5b6b224e0229a