Fix compilation on OS X.
This patch fix several compilation issues on OS X. The most notable
changes are the following.

connectx() is a standard library function on OS X, as a consequence we
must overloaded it using a macro-based trick in order to use our version
of the function.

mremap() is a linux-specific system call. As a consequence, we must use
manually remap the file on other OSes.

Some format strings have been fixed since they did suppose the types of
the parameters to be known, while the standards don't fix them (the most
notable one being struct timeval.tv_usec).

Change-Id: I987876009a3e3c65cee77764e8538262b9b35ac9
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>