ichannel: fix misuse of CMSG_DATA.
On Linux, CMSG_DATA() is implemented as an ubyte[] (or equivalent) at the
end of struct cmsghdr. On OS X, the field cmsg_data is not really defined
in the structure, and instead, the macro infers the pointer from the
beginning of the structure:

 /* given pointer to struct cmsghdr, return pointer to data */
 #define CMSG_DATA(cmsg)  ((unsigned char *)(cmsg) +  __DARWIN_ALIGN32(sizeof(struct cmsghdr)))

Looks like this is done in order to guarantee the alignement of the
cmsg_data field on 32bits (which is nice!). But as a consequence, the
CMSG_DATA() is not a field but a computed value and so taking is pointer
is not even useless, but also invalid.

Moreover, even on Linux, we take a pointer to the ubyte[] field, which is
a bit inelegant. This works because &ubyte[] returns a pointer to the
storage of the array (which is a pointer to the content of the array and
thus its first element) is equivalent to casting the ubyte[] to a ubyte *
(which is an implicit cast that often causes programmers to make the
confusion between the two types... which are not the same unless used as
the parameter of a function...). The CMSG_DATA() returns a pointer to the
data and thus we should not try to dereference its returned value.

Change-Id: I8569531189830e1e3a6c860d4c5c74a204326b1e
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>