Build: add support for OS X.11.
For some reason, the built-in "echo" command on make on OS X does not seem
to support the -n flag and instead emit a spurious -n in its output,
completely breaking the build system.

Also, take into account the fact the default compiler is clang 3.7 (or
something very similar to clang 3.7... it says "Apple LLVM version 7.3.0
(clang-703.0.29)").

Change-Id: I4ea67f83c298a1068945b0fa36c36827c5e735d9
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>