zchk-time: fix tests on OS X.
mktime() specification is quite abstract, speaking of calendar time
without any additional precision. As a consequence, the implementer is
free to support the range of date he choses... and on OS X, mktime() seems
only to support dates *after* epoch. As a consequence, provides a naive
implementation of the year-day extrapolation when mktime does not do it by
itself.

Change-Id: I710d6f7af68338e4ab7f7be3fa7e0f9589c6c82a
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>