iop-json: validate @include() syntax before parsing file.
We previously read the content of the included file as soon as
@include("/path/to/file" was read, without validating this was followed by
a closing parent. The closing parent was parsed only when the included
file content has been parsed. This patch changes that behavior so that we
ensure the @include() is well-formed before reading the included file. The
main impact is that in case of a buggy include, we will receives the error
about the ill-formatted include instead of the error that depends on the
included file.

Our tests, in particular, tried to include /proc/self/cmdline, and, in
case we ran on OS X (or other non-Linux OS I guess), the reported error in
the tests was about the missing file instead of the ill-formatted include.

Change-Id: I2315140002b07815c2642a97ea1e6fb986ccf87e
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>