z: fix undefined behavior in value display.
We compute the number of digits to display in case of assertion failure
using a double bsr. The first one to identify the maximum of significative
bits in both values, the second to deduce the number of digits to display
from this.

However, in case the values where both 0 or 1, the first bsr returned 0,
causing the second one to be undefined since there bsr(0) is undefined
(and even illegal). As a consequence, depending on the compiler and the
optmization level, the number of digits to display could be either 0 or
any other value... in my particular case, it was 2 billions of digits...
which is a bit to huge for me.

Change-Id: Ie7212b5ae7c63063beec547cde83e7b0e7ecf6b9
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>