el: have an experimental implement of el_fs_watch using FSEvents.
The "fs watch" model is too tightly coupled with inotify to be portable.
This patch at least provides an implementation that compiles, more work
remains to be done in order to have a working fs_watch on OS X.

OS X provides two distinct mechanism to have watches on the filesystem.
The first one is built into BSD's kqueue and allow watching metadata changes on
files and directories. This however does not provide any mechanism to be
notified of the creating removal of a file without keeping by hand the
listing of the directory (and relisting it on each notification).

The second mechanism, specific to OS X, is the FSEvents. There is a
service (understand that there is a daemon called fseventd whose job is to
monitor the filesystem) that watches system-wide every file-system changes
in order to feed some native services of the OS (automatic backups, search
engine indexation, ...). Any program can subscribe to that service using
the FSEvent API [1]. The provides the ability to be notified of any change
in a directory and in children. The notification can be either at
directory level "something changed in the directory, but I won't say
what", or file-specific "file has been renamed", "file has been created",
"file has been removed", ... This API is not as fine-grained as inotify
(you cannot be notified of specific operations such as file being opened,
closed, ...).  Moreover, the FS Events API lacks some directional metadata
on moves, on a rename, both the source and the destination receive are
advertised (in two separate notifications) without a way to make a
distinction between the old and the new naming without running stat.

[1] https://developer.apple.com/library/mac/documentation/Darwin/Reference/FSEvents_Ref/

Change-Id: I5cae8bebbdad16d8514b6e10a17169de61c426b1
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>