ichannel: unify stream/packet implementations.
We used to have two separate ic_write implementation: one for unix
sockets, one for other sockets. This was done for two reasons: on unix
sockets, we need the ability to send file descriptors, and for seqpacket
implementation we have to manually cut the packets since this is not done
by the protocol.

However, on OS X, unix sockets don't support SEQPACKET, and we must use a
SOCK_STREAM protocole instead. This forces us to use the same control-flow
mechanisms for unix sockets as we used for non-unix sockets, such as
supporting partially sent buffers, while still needing to support sending
file descriptors.

As a consequence, unify both implementations in a single ic_write()
function that support both unix/non-unix, stream/seqpacket sockets. To do
this, we split the flag "is_stream" into a "is_seqpacket" and a "is_unix"
flags, so that we apply fd sending only on the unix channel, while
applying manual splitting on seqpacket sockets.

The fd sending code have been cleaned up in order to properly handle
alignement constraints: fds must be aligned on 4 bytes, while the approach
based on a char[] only guaranteed a 1-byte alignement, causing potential
EINVAL.

Also, on the reading side, we fix a bug in case a message attached to a fd
was partially received. Since the fd was local to ic_read(), in case the
message was in two packets, the fd might be lost when decoding the packet
(which only happens if the message the precedes a message containing a fd
almost fill the IC_PKT_MAX window).

Change-Id: I3ea2241ffef5c3aba38208320d31513f0ad46d8f
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>