container-qhash/qm: allow to use a const pointer to the key for getters
Before this patch it wasn't possible to use a const pointer as an argument
for qm_get* functions. __QM_ADD macro is split into __QM_ADD and __QM_FIND
in order to differentiate functions that only need a const pointer from
functions that actually really need a non const pointer.

Change-Id: Ie552d3b1dbb5bc51623afac2b3566c6e59972d38