thr: fix race at thread-shutdown.
When stopping a thread, the thread-local variables are deleted. For
variables created using pthread_key_create(), this is done by calling the
provided destructor function if the current value associated to the key is
non-null. For __thread variables, this is done by the runtime.

On linux, either the runtime never wipe out the __thread variables, or it
does so after the destruction of the pthread_key variables, causing
pthread_key destructor to be able to access thread-local variables.
However, on OS X, the __thread variable are cleared (memset(0)) before the
execution of the pthread_key destructors. As a consequence, the __thread
local variable are not accessible in the destructors.

By construction, we use the thr_hooks() mechanism as a mean to register
destructors for __thread variables. The thr_hooks() uses a pthread_key()
internally in order to ensure the intialization is done on all threads
(and is done only once). As a consequence, on OS X, this mechanism is
broken. To fix this, we now explicitly call thr_detach() in the main
thread function wrapper, so that we are sure the destruction occurs before
the death of the thread and thus the __thread variables remain accessible.

Change-Id: Ifea70606680f7706f858452ebc4e514f035b5fd8
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>