el: always take "high"-res clock.
Not taking the high-res clock is dangerous since we have absolutely no
idea when the last clock tick was cached. As a consequence, in case the
event loop get stuck (for some probably *bad* reason), calling
el_timer_register() would always based itself on the beginning of the last
event loop, which may be minutes ago, causing the timer to be fired too
soon.

This was observed on LMS where a timer of 3600 was fired after less than
half an hour because the event loop is stuck in the refreshing of the
exposure and of the targets.

Change-Id: I36fcb641f6e41c672e4ed6978ff711c0ac5b9b09
Closes: #44948
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>