iop-json: rework c_to_camelcase()
The function gets a new signature:
* It takes a sb as third argument that will receive the result.
* It returns a int equals to 0 when succeeding and -1 otherwise.

This way the caller is responsible for the allocation.

This patch also removes the 2 t_scope introduced in 323677bf.
These t_scope were introducing use-after-poison errors.
Since the JSON parser is allocated on the t_stack most of unpacking
function in iop-json.blk are "secretly" t_functions, we can actually
safely (and must) remove these 2 t_scope.

Refs: #42985
Change-Id: I27c5d56dd29adb815de95ec50946bfbcb7a08ccf