el: implements a "tolerance" mechanism for LOWRES timers.
Since 1e8ece1245bc98d9ec10047b5ba508017cd4b50a the LOWRES mechanism has
been completely deactivated since it could cause timers to be fired too
early because the LOWRES applied to the time evaluation at the creation of
the timer.

We now introduce a mechanism for LOWRES timers that is based on the
expiration date instead of the creation date. The expiration of a timer is
now a range [expiry,expiry+tolerance]. We guarantee that timers are fired
within that range of time (assuming the eventloop is not blocked by a
callback that don't give control back).

The event loop sorts its timers by the upper bound of the expiration
range, but executes all timers whose lower range has expired in
el_timer_process(). That way a low-res timer will be executed if it
reached if upper-bound timeout, or if another timer has expired within the
range [expiry,expiry+tolerance], causing low-res timers to be fired if
they expired and a higher-res timer expired too.

LOWRES timers are timers for which the tolerance is non-0. The tolerance
is computed as a factor of the timer duration (currently arbitrary chosen
as 10% of the timer duration) based on the assumption that the longer the
timer is, the more probable we don't care about the precision (for example
if we fire a timer every hour, we probably don't care about having a
sub-second precision. BTW, the tolerance is capped at 59 seconds (which is
the default timeout of a loop of the event loop) since we always check for
expired timers at the beginning of the event loop.

Change-Id: I83f8dc308c08b3a26c4f3c48ea776d3fe74480d1
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>