qpage: MADV_DONTNEED is not a portable way to clear memory.
The posix documentation for MADV_DONTNEED only says it's a way to say the
kernel we won't need the pages anytime soon. On Linux, for anonymous
memory, this is equivalent to clearing memory, however this is not the
case on OS X.

A portable way to efficiently clear pages is to remap them.

Note that this was responsible for inconsitent behavior in qpage that
completely broke QDB on OS X.

Change-Id: Ib1b03b78305dfdfdd40f1657d7cd54686c1cd1be
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>