iop-jpack: support packing the first field of a struct in a subfile.
Before, it was not possible to pack the first field of a  structure in a
sub-file, because its pointer is the same as the pointer of the
structure itself in the parent object (this is the structure that would
be packed instead).

Now, we use iop paths instead of pointers.

The structures iop_junpack_subfile_t and iop_junpack_field_path_t are
now factorized and renamed to be used by the packer also.

Change-Id: I1454ee55516e75432264ad633d364e5298d84122
Refs: #44738