core-obj: allow to pass additional params at declaration
Let OBJ_CLASS and OBJ_CLASS_NO_TYPEDEF take additional parameters to pass
to the FIELDS and METHODS macros. Can be used to specify the type of a
pointer in order to avoid casts.

Warning ! This feature should be used carefully since the user has to make
sure that the size of the fields generated by a given FIELDS macro should
*never* change.

Change-Id: I9aad291bb93d177e05b86660da010a7b1438d5f5