dlist: get rid of non-safe iterators
In the worst cases of my benches I get a dlist_for_each_safe() 4% slower
than the non-safe version. In more real usage of dlist_for_each_safe()
I get only 0.4% of difference…

Thus it doesn't seem legitimate to keep a non-safe version of
dlist_for_each.

However be careful, there is a side-effect of this patch. With the
non-safe version of dlist_for_each(), a dlist_add() right after the
current item would have made it the next item in the iteration and now we
will ignore it.

Change-Id: Ic8e986ba8232d810c6f0837e4feb881456cd6f76