iop[rpc]: don't give volatile memory to RPC implementation
When we send an ic_msg_t to a local IChannel (ic_is_local) and in the case
we don't have the original IOP object, we parse the payload stored in the
ic_msg_t. With the previous implementation it meant we gave an unpacked
IOP object (arg) to the RPC implementation holding references on the
payload of the ic_msg_t. Since the ic_msg_t is deleted when calling
ic_reply() it left us with invalid memory for the rest of the RPC
implementation execution.

The solution is to simply ask the IOP unpacker to duplicate memory instead
of taking references to the payload in this case.

Change-Id: I33ff7f04bfe4387e86a627443626265f7b8a0e23