iop: fix heterogeneity between handling of types DATA and STRING or XML
Once upon a time, the C type for IOP_T_DATA was iop_data_t and not lstr_t.
It seems that at the time of the migration to lstr_t, some adaptation had
been forgotten.

This fixes a potential bug in __iop_copy(): the memory pool flags were not
set for type DATA.

Change-Id: I2a8790ba89ae1ef29b7a2f82add85e2d542b6f83
Signed-off-by: Sébastien Thouément <sebastien.thouement@intersec.com>