el: Rework ev allocation.
The final goal of this patch is to get rid of the huge static array
el_g.evs_initial, which takes 64k in every binary.

In order to do that, we get rid of it and only use dynamically-allocated
buckets. Also rework the allocation code to make it a bit clearer.

The big question is now: "can we delete our buckets at module shutdown?".
The answer is no if there are still events in use at this time (which
should not happen, but sadly happens in almost all our binaries).

Note that with the previous code, the allocated buckets were always
leaked...

Change-Id: I785671f050eb89e79b5600f6f05b816d520e57bf