file-log: Do not el_unref children, wait for gzip termination in tests.
The test of file_log did not leave the el module in a clean state because
it did not wait for the termination of the forked gzip processes, which
caused "still reachable" traces under valgrind.

We now call el_loop at the end of the tests so that we wait for the
children. This required to removing the el_unref of the el_child, which
looked quite dangerous anyway.

Change-Id: I5c73cfdea9ba5788ad0c070dfeee494186207f21