iop-dso: Fix dso_close order issue.
The goal of this patch is to fix once for all the fact that when a dso A
uses symbols defined in a dso B (because of fixups), A must be closes
before B (because the orphan classes check blows otherwise).

This is achieved by tracking, for each dso, the list of dso using it and
used by it.

When closing a dso, the other dsos using it are properly re-loaded, which
is a way to implement the "unfixup".

Change-Id: Ie4408de4c9984b99dd7eca8d5d9e9a22ba0e8752
Closes: #46008