Build: rewrite _list_checks in python
The previous script was hard to read and therefore was hard to maintain
too.

This patch does the same thing that before and it adds the ability
to add some new features in the next patch like add tags in ZFile. It
works on python 2.4 to 3.X.

Also, this patch backports the output behavior from origin/2014.4. To
avoid several patch to fix all branches.

Change-Id: Iae2f4cd35776fbe0e5ab59fa27a87c60bd290c57
Refs: #45873