iop-dso: fix bug in fixup dependency tracking
When loading the "same" dso twice, the second one was fixed up on the
first one, which it should not, causing issues when unloading the first
one if the file of the second dso was removed.

Fix this by avoiding a fixup if the pkg found for a struct is our own.

Change-Id: I2b213af23735dca24c2f01c2137c897f66dac27c