httpc: allow absolute URL in queries
Handling of URI in httpc_query_start_flags() was wrong when used with
absolute URIs and proxies. In case of proxy, we mustn't try to transform
the URI into an absolute URL if the URL is already absolute.

(cherry picked from commit 12b5eeb41740b786b65f23a29382950af428b647)

Refs: #46613
Change-Id: I379422f43d4c34bf6bea15e10bab636d32108e89