Build: add tags in ZFile
This patch adds the ability to include tags in ZFile. The goal is to tags
to slow some tests suites. All tags are prefixed with a '@'. If one or
more tags are in Z_TAG_SKIP the test suite is skipped.

Here some examples of usage:
zchk  @perf, @upgrade
zchk_python.py  # @slow
testem/backbone-ic-testem.json @wip

Change-Id: I82de67a2fa651a60f136f520cf49a5684f270366