qhash: introduce qh/qm_khptr
Those new qh/qm templates provide a replacement for cases where we used
hash table of a pointer type with qhash_hash_ptr/qhash_ptr_equal as
callbacks.

Instead of using a hashing and an equality function, those new template
are based on the fact pointers are basically 64bit integer and provide use
the qh_k64 backend which is much faster than the generic backend.

As a consequence, using a qm_khptr is 10% faster than using a qm_kptr with
the standard pointer hashing functions and avoid having to define a
specific hashing/equality function.

Change-Id: Ib54edc570f8d3d1a4637d2b78cc8d51470a0c039
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>