el: Warn about events still in use in el_shutdown.
This trace will allow to fix our code, and maybe one day assert it does
not happen...

Problem: when logger_is_traced is evaluated, we are in el_shutdown and the
log module is already shutdown, so if this is the first time we evaluate
the logger level, IS_DEBUG is not taken into account. This will be
workarounded in the next commit.

Change-Id: I6cf4a641fdab8f6d50549f929af53652ccaf3c1b