build: www-check is behave type
commit 6724aedc741 has changed the Build/_list_checks.sh to
Build/_list_checks.py but has been added to 2013.3 where the Z_LIST_SKIP
didn't even exist which is awkward.
Z_LIST_SKIP has been introduced in 2014.4 by 2445711b95253f76 but it
behaved differently back then as "behave" was treated as "C" category
by _list_checks.sh
As the the new _list_checks.py treat "behave" as "behave", behave is now
launched when it should be skipped between 2014.4 and 2015.1.
It's only in 2015.3 with 777b9076 that "behave" had been treated as
"behave" category.

The easiest way to not bother about merge problems is to use the same
syntax as in 777b9076 so we just add "behave" to the Z_LIST_SKIP

Change-Id: Ia46fdad773bd2e4f55427f64fb0284c0c4634a84