iopc-dso: Launch gcc in its own process group id.
In platform, I've proposed a patch (Ic610af2403) whose one consequence is
that every fork of a process receives a termination signal when the master
tries to stop a process.

This causes an issue if we try to stop the core while it is building its
event.so, because gcc terminates with an error, and the core then commits
suicide:

 https://git.corp/buildbot/builders/review.centos6+/builds/28215

A solution is to create a dedicated process group of the forks of gcc, so
that then end normally.  GCC is stable enough so that we know it won't
hang for hours, so this should not be a problem.

Change-Id: I23ce6872692472548e60b28a109dbf379792a16b
Refs: #47223