iop-dso: iopdso_fix_class_parent should not fix the ancestors.
This function, whose purpose is to fixup the "parent" pointer of a class,
was doing it recursively, which was a mistake because we could fixup
pointers of a package without the "use_external_packages" flag!

This is what caused the bug described in #47104:
* load igloo.so,
* load one.so: no fixup at this point, since one.so does not use external
  packages.
* load one/event.iop: the parent of the class event.BonusCredit,
  qre.base.Bonus, is fixed-up to the version of one.so, and then the
  parent of the qre.base.Bonus of one.so, bigdata.base.Event, is fixed-up
  to the version of igloo.so, which is clearly wrong.

These is no reason to do this recursively: all the classes of the package
being registered will be fixed-up anyway.

Change-Id: Icebf3d0862a5aba0d296af8f840c29d9c4d708e1
Closes: #47104