asn1-aper: fix the "bit-field case" encoding/decoding for size == 8
See ITU-T X.691-0207 § 10.5.7.1: when encoding/decoding fully constrained
numbers with range between 129 and 255 (encoding size == 8), we used to
apply the rule described in § 10.5.7.2 (The one-octet case), that forces
the alignment.

Warning: the value of "range" in the specification is max - min + 1, while
the value we actually use in the code is "d_max" (max - min).

The bug was seen because of bit strings described this way:

    BIT STRING (SIZE (1..160, ...))

The value of the "range" for the encoding of the length is 160,
in [129, 255], and the extension bit to encode first makes us encounter
the bug easily because in most of the cases we don't begin the writing of
the length while aligned, so the missing alignment is visible.

Change-Id: I231b7b305621f0913000cf8a0dc918d715d7f16a