core-types: introduce OPT_COPY
Since in IOP the optional fields are generated with anonymous structures
(for optional structures, not for optional basic types, like integers),
it isn't possible to do 'opt_a = opt_b;' with these fields.
To improve the readability of the code that has to do such copies, let's
introduce a macro doing it.

Change-Id: I74b248c3fd7aa58fc229c683a6cb5ee2b187ebcf