el: Call module_on_term directly in el_signal_process.
Before this patch, it was up to each program to call module_on_term in
their on_term function. But in practice, many program don't do that
(because this is a quire recent feature), causing bugs like the one fixed
by https://git.corp/r/#/c/71626/.

Note that this is still necessary to register the signals to that
module_on_term is called.

Change-Id: Ided011f86d0f7aca5a132d51f45a922fa31879ff