el: Prevent leak of FS_WATCH events.
FS_WATCH events are actually not destroyed when they are unregistered, but
when we have confirmation that inotify took it into account.
To know that, the el library registers a inotify FD event and reads on it.
This event is unregistered when we have confirmation that no path is
watched by inotify anymore.

As a consequence, the processes that were unregistering their FS_WATCH in
their shutdown were actually leaking them, plus the internal event of the
library.

Fix that by:
* not unref'ing the internal event, so that the event loop does not exit
  before all FS_WATCH are properly deleted.
* forbid to unref a FS_WATCH, because this would clearly be buggy.

Change-Id: I59f38273661fbd7cabac91db1ce47eb92248b443