asn1-aper: fix potential invalid read in aper_encode_choice
The choice index is not necessarily written in an "int", in fact, it can
be written on an "uint16_t" if the choice is adapted from an IOP union for
example.

This change should have been done in
204a0ef9ecf189f17228f800f8fa1401ae50f82f, that adds compatibility helpers
between IOP unions and ASN.1 choices, but adaptations in ASN.1 PER had
been forgotten.

Change-Id: Id24d42b78e5649b5434f8066fab438aa510cd027